requirements.txt
keras==2.6.0
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.1
plotly==5.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
streamlit==0.88.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
